pyproject.toml
[tool.poetry]
name = "bank-api"
version = "0.1.0"
description = ""
authors = ["Eduardo Victor <eduvictornobrega@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
uvloop = "^0.21.0"
httptools = "^0.6.4"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
pre-commit = "^4.1.0"
httpx = "^0.28.1"
asgi-lifespan = "^2.1.0"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.12"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
pythonpath=["./src"]
testpaths = ["./app"]
markers = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
